+++ lib/js_of_ocaml/dom_html.mli
@@ -29,6 +29,15 @@ open Js
 
 class type cssStyleDeclaration =
   object
+    method setProperty :
+      js_string t -> js_string t -> js_string t optdef -> js_string t meth
+
+    method getPropertyValue : js_string t -> js_string t meth
+
+    method getPropertyPriority : js_string t -> js_string t meth
+
+    method removeProperty : js_string t -> js_string t meth
+
     method animation : js_string t prop
 
     method animationDelay : js_string t prop
@@ -629,6 +638,15 @@ and animationEvent =
 and mediaEvent =
   object
     inherit event
+  end
+
+and messageEvent =
+  object
+    inherit event
+
+    method data : Unsafe.any opt readonly_prop
+
+    method source : Unsafe.any opt readonly_prop
   end
 
 (** {2 HTML elements} *)
@@ -2439,6 +2457,8 @@ module Event : sig
 
   val lostpointercapture : pointerEvent t typ
 
+  val message : messageEvent t typ
+
   val pause : mediaEvent t typ
 
   val play : mediaEvent t typ
@@ -2972,6 +2992,7 @@ val opt_tagged : #element t opt -> taggedElement option
 type taggedEvent =
   | MouseEvent of mouseEvent t
   | KeyboardEvent of keyboardEvent t
+  | MessageEvent of messageEvent t
   | MousewheelEvent of mousewheelEvent t
   | MouseScrollEvent of mouseScrollEvent t
   | PopStateEvent of popStateEvent t
@@ -3123,6 +3144,8 @@ module CoerceTo : sig
   val mouseScrollEvent : #event t -> mouseScrollEvent t opt
 
   val popStateEvent : #event t -> popStateEvent t opt
+
+  val messageEvent : #event t -> messageEvent t opt
 end
 
 type timeout_id_safe